+++ continuous_integration/travis/install.sh
@@ -28,6 +28,7 @@ source activate test-environment
 
 # Install dependencies
 conda install -q \
+    asyncssh \
     bokeh \
     click \
     coverage \
