+++ distributed/deploy/__init__.py
@@ -2,6 +2,7 @@ from ..utils import ignoring
 
 from .cluster import Cluster
 from .local import LocalCluster
+from .ssh import SSHCluster
 from .spec import SpecCluster, ProcessInterface
 from .adaptive import Adaptive
